More hacking on getting this working.

+++ library.go
@@ -3,6 +3,7 @@ package kubeboard
 import (
 	"os"
 	"os/exec"
+	"syscall"
 
 	"github.com/pieterclaerhout/go-log"
 	"github.com/zserge/webview"
@@ -18,17 +19,22 @@ func NewKubeBoard() KubeBoard {
 
 func (kb KubeBoard) Start() error {
 	go kb.startProxy()
-	return kb.startWebUI()
+	kb.startWebUI()
+	return nil
 }
 
 func (kb KubeBoard) Stop() {
-	kb.kubectlCmd.Process.Kill()
-
+	log.Info("Closing down")
+	if kb.kubectlCmd != nil && kb.kubectlCmd.Process != nil {
+		kb.kubectlCmd.Process.Kill()
+	}
 }
 
 func (kb KubeBoard) startProxy() {
 
 	log.Info("Starting kubectl proxy")
+
+	// TODO: kill the previous instance if any
 
 	file, _ := os.Create("/Users/pclaerhout/Desktop/kubectl.log")
 	defer file.Close()
@@ -39,19 +45,34 @@ func (kb KubeBoard) startProxy() {
 	kb.kubectlCmd = exec.Command("/usr/local/bin/kubectl", "proxy", "--port=8001")
 	kb.kubectlCmd.Stdout = file
 	kb.kubectlCmd.Stderr = file
+	kb.kubectlCmd.SysProcAttr = &syscall.SysProcAttr{Setpgid: true}
 	if err := kb.kubectlCmd.Start(); err != nil {
 		file.WriteString(err.Error() + "\n")
 	}
 
 }
 
-func (kb KubeBoard) startWebUI() error {
+func (kb KubeBoard) startWebUI() {
 	log.Info("Starting web UI")
-	return webview.Open(
-		"kubeboard",
-		"http://localhost:8001/api/v1/namespaces/kube-system/services/https:kubernetes-dashboard:/proxy/#!/login",
-		1280,
-		960,
-		true,
-	)
+
+	wv := webview.New(webview.Settings{
+		Title:     "kubeboard",
+		URL:       "http://localhost:8001/api/v1/namespaces/kube-system/services/https:kubernetes-dashboard:/proxy/#!/login",
+		Width:     1280,
+		Height:    960,
+		Resizable: true,
+	})
+	defer func() {
+		kb.Stop()
+		wv.Exit()
+	}()
+	wv.Run()
+
+	// return webview.Open(
+	// 	"kubeboard",
+	// 	"http://localhost:8001/api/v1/namespaces/kube-system/services/https:kubernetes-dashboard:/proxy/#!/login",
+	// 	1280,
+	// 	960,
+	// 	true,
+	// )
 }
